pyproject.toml
[project]
name = "moldet-mcp-server"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "jibsn", email = "jibsn@outlook.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "absl-py==2.3.1",
    "albumentations==1.1.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "authlib==1.6.3",
    "blinker==1.9.0",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "configargparse==1.7.1",
    "contourpy==1.3.2",
    "cryptography==45.0.7",
    "cycler==0.12.1",
    "cyclopts==3.24.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "docutils==0.22",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.0",
    "fastmcp==2.12.3",
    "fastprogress==1.0.3",
    "flask==3.1.2",
    "fonttools==4.59.2",
    "gensim==4.3.2",
    "grpcio==1.74.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "imageio==2.37.0",
    "isodate==0.7.2",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jsonschema==4.25.1",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.9.1",
    "kiwisolver==1.4.9",
    "lazy-loader==0.4",
    "lazy-object-proxy==1.12.0",
    "markdown==3.9",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.6",
    "mcp==1.14.0",
    "mdurl==0.1.2",
    "molscribe==1.1.1",
    "more-itertools==10.8.0",
    "networkx==3.4.2",
    "numpy==1.26.4",
    "nvidia-cublas-cu11==11.10.3.66",
    "nvidia-cuda-nvrtc-cu11==11.7.99",
    "nvidia-cuda-runtime-cu11==11.7.99",
    "nvidia-cudnn-cu11==8.5.0.96",
    "openapi-core==0.19.5",
    "openapi-pydantic==0.5.1",
    "openapi-schema-validator==0.6.3",
    "openapi-spec-validator==0.7.2",
    "opencv-python==4.12.0.88",
    "opencv-python-headless==4.12.0.88",
    "opennmt-py==2.2.0",
    "packaging==25.0",
    "pandas==2.3.2",
    "parse==1.20.2",
    "pathable==0.4.4",
    "pillow==11.3.0",
    "polars==1.33.1",
    "protobuf==6.32.1",
    "psutil==7.0.0",
    "pycparser==2.23",
    "pydantic==2.11.8",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pymupdf==1.26.4",
    "pyonmttok==1.37.1",
    "pyparsing==3.2.4",
    "pyperclip==1.9.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "qudida==0.0.4",
    "rdkit-pypi==2022.9.5",
    "referencing==0.36.2",
    "requests==2.32.5",
    "rfc3339-validator==0.1.4",
    "rich==14.1.0",
    "rich-rst==1.3.1",
    "rpds-py==0.27.1",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.2",
    "scipy==1.15.3",
    "sentencepiece==0.2.1",
    "setuptools==80.9.0",
    "six==1.17.0",
    "smart-open==7.3.1",
    "smilespe==0.0.3",
    "sniffio==1.3.1",
    "sse-starlette==3.0.2",
    "starlette==0.47.3",
    "tensorboard==2.20.0",
    "tensorboard-data-server==0.7.2",
    "threadpoolctl==3.6.0",
    "tifffile==2025.5.10",
    "timm==0.4.12",
    "torch==1.13.1",
    "torchtext==0.5.0",
    "torchvision==0.14.1",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "ultralytics==8.3.199",
    "ultralytics-thop==2.0.17",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "waitress==3.0.2",
    "werkzeug==3.1.1",
    "wheel==0.45.1",
    "wrapt==1.17.3",
]
